requirements.txt
flask>=3.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=10.0.0
h5py>=3.9.0
tensorflow>=2.13.0
werkzeug>=3.0.0
